add more int tests (#175)
fix neon constructor from int args parameter pack and add more int tests

+++ include/xsimd/types/xsimd_neon_uint8.hpp
@@ -82,7 +82,7 @@ namespace xsimd
 
     template <class... Args, class>
     inline batch<uint8_t, 16>::batch(Args... args)
-        : m_value{args...}
+        : m_value{static_cast<uint8_t>(args)...}
     {
     }
